% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{AbxParams}
\alias{AbxParams}
\title{Antibiotic Parameters}
\usage{
AbxParams(onoff = FALSE, delay = 0, life = 1)
}
\arguments{
\item{onoff}{The mode of the antibiotics, dose with delay and life vs on/off events.}

\item{delay}{The delay in using antibiotics.}

\item{life}{The life of antibiotics.}
}
\description{
Antibiotic Parameters
}
\examples{
AbxParams()
}
