% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{AbxRateParams}
\alias{AbxRateParams}
\title{Antibiotic Administration Rate Parameters}
\usage{
AbxRateParams(
  uncolonized = Param(1, 0),
  colonized = Param(1, 0),
  recovered = Param(0)
)
}
\arguments{
\item{uncolonized}{Rate of antibiotic administration when the individual is uncolonized.}

\item{colonized}{Rate of antibiotic administration when the individual is colonized.}

\item{recovered}{Rate of antibiotic administration when the individual is recovered.}
}
\value{
list of parameters for antibiotic administration.
}
\description{
Antibiotic Administration Rate Parameters
}
\examples{
AbxRateParams()
}
