% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{LinearAbxAcquisitionParams}
\alias{LinearAbxAcquisitionParams}
\title{Title}
\usage{
LinearAbxAcquisitionParams(
  base = Param(0.01),
  time = Param(1, 0),
  mass = Param(1),
  freq = Param(1),
  col_abx = Param(1, 0),
  suss_abx = Param(1, 0),
  suss_ever = Param(1, 0)
)
}
\arguments{
\item{suss_ever}{}
}
\description{
Title
}
