% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{LogNormalModelParams}
\alias{LogNormalModelParams}
\alias{LinearAbxModel}
\title{Model Parameters for a Log Normal Model}
\usage{
LogNormalModelParams(
  modname,
  nstates = 2,
  nmetro = 1000,
  forward = TRUE,
  cheat = FALSE,
  Insitu = InsituParams(),
  SurveilenceTest = SurveillanceTestParams(),
  ClinicalTest = ClinicalTestParams(),
  OutOfUnitInfection = OutOfUnitInfectionParams(),
  InUnit = LogNormalInUnitAcquisition(),
  Abx = AbxParams(),
  AbxRate = AbxRateParams()
)

LinearAbxModel(..., InUnit = LogNormalABXInUnitParameters())
}
\arguments{
\item{modname}{The name of the model used. Usually specified by specification functions.}

\item{nstates}{The number of states in the model.}

\item{nmetro}{TODO}

\item{forward}{TODO}

\item{cheat}{TODO}

\item{Insitu}{In Situ Parameters}

\item{SurveilenceTest}{Surveillance Testing Parameters}

\item{ClinicalTest}{Clinical Testing Parameters}

\item{OutOfUnitInfection}{Out of Unit Infection Parameters}

\item{InUnit}{In Unit Parameters, should be a list of lists with parameters
for the acquisition, progression and clearance of the disease.}

\item{Abx}{Antibiotic Parameters}

\item{AbxRate}{Antibiotic Rate Parameters}
}
\description{
Model Parameters for a Log Normal Model
}
\section{Functions}{
\itemize{
\item \code{LinearAbxModel()}: Linear Antibiotic Model Alias

}}
