% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newCppModelInternal}
\alias{newCppModelInternal}
\title{Create a new C++ model object wrapped in appropriate reference class}
\usage{
newCppModelInternal(modelParameters, verbose = FALSE)
}
\arguments{
\item{modelParameters}{List of model parameters (same format as runMCMC)}

\item{verbose}{Print progress messages (default: false)}
}
\value{
A reference class object wrapping the C++ model:
\itemize{
\item CppLogNormalModel - for "LogNormalModel"
\item CppLinearAbxModel - for "LinearAbxModel"
\item CppLinearAbxModel2 - for "LinearAbxModel2"
\item CppMixedModel - for "MixedModel" (note: needs Module exposure)
}

The returned object provides access to model methods and properties including:
\itemize{
\item InColParams, OutColParams, InsituParams, etc.
\item logLikelihood(), getHistoryLinkLogLikelihoods(), etc.
}
}
\description{
Creates and initializes a C++ model object based on the provided parameters,
then wraps it in the appropriate R reference class that exposes the model's
methods and properties.
}
