% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{OutOfUnitInfectionParams}
\alias{OutOfUnitInfectionParams}
\title{Out of Unit Infection Parameters}
\usage{
OutOfUnitInfectionParams(
  acquisition = Param(0),
  clearance = Param(0),
  thirdoption = Param(0)
)
}
\arguments{
\item{thirdoption}{What should this be?}
}
\value{
A list of parameters for out of unit infection.
}
\description{
Out of Unit Infection Parameters
}
