% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{Param}
\alias{Param}
\title{Construct a parameter with a prior, weight and an update flag.}
\usage{
Param(
  init,
  weight = if_else(init == 0, 0, 1),
  update = weight > 0,
  prior = init
)
}
\arguments{
\item{init}{the initial value of the parameter.}

\item{weight}{the weight of the prior.}

\item{update}{a flag indicating if the parameter shouldbe updated in the MCMC.}

\item{prior}{mean value of the prior distribution, may be used with weight to fully determine prior parameters.}
}
\description{
Construct a parameter with a prior, weight and an update flag.
}
