% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{ProgressionParams}
\alias{ProgressionParams}
\title{Progression Parameters}
\usage{
ProgressionParams(
  rate = Param(0.01),
  abx = Param(1, 0),
  ever_abx = Param(1, 0)
)
}
\arguments{
\item{ever_abx}{}
}
\description{
Progression Parameters
}
