% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{SurveillanceTestParams}
\alias{SurveillanceTestParams}
\title{Surveillance Test Parameters}
\usage{
SurveillanceTestParams(
  colonized = Param(0, 0),
  uncolonized = Param(0.8, 1),
  recovered = Param(0, 0)
)
}
\arguments{
\item{colonized}{Also known as the true positive rate for a two state model.}

\item{uncolonized}{Also known as the false positive rate for a two state model.}

\item{recovered}{The rate of positive tests when the individual is in the (optional) recovered state.}
}
\description{
Specify the rates of positive tests for each state of the model.
}
