% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{SurveillanceTestParamsAbx}
\alias{SurveillanceTestParamsAbx}
\title{Surveillance Test Parameters with Antibiotics}
\usage{
SurveillanceTestParamsAbx(
  colonized_off = Param(0, 0),
  colonized_on = Param(0, 0),
  uncolonized_off = Param(0.8, 1),
  uncolonized_on = Param(0.8, 1),
  latent_off = Param(0, 0),
  latent_on = Param(0, 0)
)
}
\arguments{
\item{colonized_off, colonized_on}{The probability of receiving a positive test when colonized, without or with antibiotic use, respectively.}

\item{uncolonized_off, uncolonized_on}{The probability of receiving a positive test when uncolonized, without or with antibiotic use, respectively.}

\item{latent_off, latent_on}{The probability of receiving a positive test when in latent state, without or with antibiotic use, respectively.}
}
\description{
Specify the rates of positive tests for each state of the model.
}
\examples{
SurveillanceTestParamsAbx()
}
