% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{testHistoryLinkLogLikelihoods}
\alias{testHistoryLinkLogLikelihoods}
\title{Get Individual HistoryLink Log Likelihoods (Diagnostic Function)}
\usage{
testHistoryLinkLogLikelihoods(modelParameters)
}
\arguments{
\item{modelParameters}{List of model parameters (same format as runMCMC)}
}
\value{
List containing:
\itemize{
\item \code{linkLogLikelihoods} - vector of log likelihoods for each link
\item \code{overallLogLikelihood} - total log likelihood from model->logLikelihood()
\item \code{numLinks} - number of history links
}
}
\description{
This function creates a model and system history, then returns the log likelihood
contribution from each individual HistoryLink in the system. Useful for debugging
and verifying likelihood calculations.
}
