% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\docType{data}
\name{InsituParams}
\alias{InsituParams}
\title{Title}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
InsituParams
}
\arguments{
\item{doit}{}
}
\description{
Title
}
\keyword{datasets}
