% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runMCMC}
\alias{runMCMC}
\title{Run Bayesian Transmission MCMC}
\usage{
runMCMC(data, MCMCParameters, modelParameters, verbose = FALSE)
}
\arguments{
\item{data}{Data frame with columns, in order: facility, unit, time, patient, and event type.}

\item{MCMCParameters}{List of MCMC parameters.}

\item{modelParameters}{List of model parameters, see \if{html}{\out{<LogNormalModelParams>}}.}

\item{verbose}{Print progress messages.}
}
\value{
A list with the following elements:
\itemize{
\item \code{Parameters} the MCMC chain of model parameters
\item \code{LogLikelihood} the log likelihood of the model at each iteration
\item \code{MCMCParameters} the MCMC parameters used
\item \code{ModelParameters} the model parameters used
\item \code{ModelName} the name of the model
\item \code{nstates} the number of states in the model
\item \code{waic1} the WAIC1 estimate
\item \code{waic2} the WAIC2 estimate
\item and optionally (if \code{MCMCParameters$outputfinal} is true) \code{FinalModel} the final model state.
}
}
\description{
Run Bayesian Transmission MCMC
}
